+++ src/types/global.d.ts
@@ -13,3 +13,18 @@ declare module '@kappasigmamu/canary-component'
 declare class ThreeCanary {
   constructor(objectUrl?: string)
 }
+
+declare interface SocietyMember {
+  accountId: AccountId
+  hasPayouts: boolean
+  hasStrikes: boolean
+  isDefenderVoter: boolean
+  isFounder: boolean
+  isHead: boolean
+  isSkeptic: boolean
+  isSuspended: boolean
+  isWarned: boolean
+  payouts: [BlockNumber, Balance][]
+  strikes: StrikeCount
+  strikesCount: number
+}
